LayoutTests/fast/dom/webtiming-navigate-within-document-expected.txt

This test checks that navigating within the document does not reset Web Timing numbers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS requestEnd is unchanged.
PASS responseStart is unchanged.
PASS connectStart is unchanged.
PASS domainLookupStart is unchanged.
PASS connectEnd is unchanged.
PASS responseEnd is unchanged.
PASS requestStart is unchanged.
PASS navigationStart is unchanged.
PASS loadEventEnd is unchanged.
PASS redirectStart is unchanged.
PASS domainLookupEnd is unchanged.
PASS unloadEventEnd is unchanged.
PASS fetchStart is unchanged.
PASS loadEventStart is unchanged.
PASS redirectEnd is unchanged.
PASS successfullyParsed is true

TEST COMPLETE
